refactor: split forge into a shared core and a GitHub backend module
Move the GitHub-shaped code — gh invocation, GraphQL query build/parse,
and JSON normalization — into src/forge/github.rs. src/forge.rs keeps the
forge-agnostic contract: the PrView/PrSnapshot model, the resolution and
pick logic, sync derivation, and a subprocess runner parameterized by
program so a second forge CLI can reuse it. No behavior change; tests
move with their subjects.
